feat(collapsibles) - wip transforms.ts
#172 wip
