+++ package.json
@@ -78,6 +78,7 @@
 	},
 	"dependencies": {
 		"@wordpress/a11y": "^4.10.0",
+		"@wordpress/shortcode": "^4.12.0",
 		"color": "^4.2.3",
 		"gsap": "^3.12.5",
 		"lodash": "^4.17.21",
